+++ src/app/pages/portal/portal.component.ts
@@ -7,8 +7,8 @@ import {
   ElementRef
 } from '@angular/core';
 import { ActivatedRoute, Params } from '@angular/router';
-import { Subscription, of, BehaviorSubject, combineLatest } from 'rxjs';
-import { debounceTime, take, pairwise, skipWhile, first } from 'rxjs/operators';
+import { Subscription, of, BehaviorSubject, combineLatest, zip } from 'rxjs';
+import { debounceTime, take, pairwise, skipWhile, first, concatMap, delay } from 'rxjs/operators';
 import { MatDialog, MatDialogConfig, MatDialogRef } from '@angular/material/dialog';
 import MapBrowserEvent from 'ol/MapBrowserEvent';
 import * as olProj from 'ol/proj';
@@ -68,7 +68,10 @@ import {
   addStopToStore,
   ImageLayer,
   VectorLayer,
-  MapExtent
+  MapExtent,
+  moveToOlFeatures,
+  FeatureMotion,
+  ConfigFileToGeoDBService
 } from '@igo2/geo';
 
 import {
@@ -97,6 +100,7 @@ import { WelcomeWindowService } from './welcome-window/welcome-window.service';
 import { MatPaginator } from '@angular/material/paginator';
 import { ObjectUtils } from '@igo2/utils';
 import olFormatGeoJSON from 'ol/format/GeoJSON';
+import { PwaService } from '../../services/pwa.service';
 
 @Component({
   selector: 'app-portal',
@@ -235,6 +239,11 @@ export class PortalComponent implements OnInit, OnDestroy {
   }
   set expansionPanelExpanded(value: boolean) {
     this.workspaceState.workspacePanelExpanded = value;
+    if (value === true) {
+      this.map.viewController.setPadding({bottom: 280});
+    } else {
+      this.map.viewController.setPadding({bottom: 0});
+    }
   }
 
   get toastPanelShown(): boolean {
@@ -323,7 +332,9 @@ export class PortalComponent implements OnInit, OnDestroy {
     private queryService: QueryService,
     private storageService: StorageService,
     private editionWorkspaceService: EditionWorkspaceService,
-    private directionState: DirectionState
+    private directionState: DirectionState,
+    private pwaService: PwaService,
+    private configFileToGeoDBService: ConfigFileToGeoDBService
   ) {
     this.hasExpansionPanel = this.configService.getConfig('hasExpansionPanel');
     this.showSimpleFilters = this.configService.getConfig('simpleFilters') === undefined ? false : true;
@@ -336,8 +347,8 @@ export class PortalComponent implements OnInit, OnDestroy {
       this.configService.getConfig('showRotationButtonIfNoRotation');
     this.showMenuButton = this.configService.getConfig('showMenuButton') === undefined ? true :
       this.configService.getConfig('showMenuButton');
-    this.showSearchBar = this.configService.getConfig('showSearchBar') === undefined ? true :
-      this.configService.getConfig('showSearchBar');
+    this.showSearchBar = this.configService.getConfig('searchBar.showSearchBar') === undefined ? true :
+      this.configService.getConfig('searchBar.showSearchBar');
     this.forceCoordsNA = this.configService.getConfig('app.forceCoordsNA');
     this.hasFeatureEmphasisOnSelection = this.configService.getConfig('hasFeatureEmphasisOnSelection');
     this.igoSearchPointerSummaryEnabled = this.configService.getConfig('hasSearchPointerSummary');
@@ -352,6 +363,10 @@ export class PortalComponent implements OnInit, OnDestroy {
 
     this.initWelcomeWindow();
 
+    this.route.queryParams.subscribe((params) => {
+      this.readLanguageParam(params);
+    });
+
     this.authService.authenticate$.subscribe((authenticated) => {
       this.contextLoaded = false;
     });
@@ -360,23 +375,23 @@ export class PortalComponent implements OnInit, OnDestroy {
       (context: DetailedContext) => this.onChangeContext(context)
     );
 
-    this.contextMenuStore.load([
-      {
-        id: 'coordinates',
-        title: 'coordinates',
-        handler: () => this.searchCoordinate(this.contextMenuCoord)
-      },
-      {
-        id: 'googleMaps',
-        title: 'googleMap',
-        handler: () => this.openGoogleMaps(this.contextMenuCoord)
-      },
-      {
-        id: 'googleStreetView',
-        title: 'googleStreetView',
-        handler: () => this.openGoogleStreetView(this.contextMenuCoord)
-      }
-    ]);
+    const contextActions = [{
+      id: 'coordinates',
+      title: 'coordinates',
+      handler: () => this.searchCoordinate(this.contextMenuCoord)
+    },
+    {
+      id: 'googleMaps',
+      title: 'googleMap',
+      handler: () => this.openGoogleMaps(this.contextMenuCoord)
+    },
+    {
+      id: 'googleStreetView',
+      title: 'googleStreetView',
+      handler: () => this.openGoogleStreetView(this.contextMenuCoord)
+    }];
+
+    this.contextMenuStore.load(contextActions);
 
     this.queryStore.count$
       .pipe(pairwise())
@@ -396,6 +411,9 @@ export class PortalComponent implements OnInit, OnDestroy {
       });
     this.map.ol.once('rendercomplete', () => {
       this.readQueryParams();
+      if (this.configService.getConfig('geolocate.activateDefault') !== undefined) {
+        this.map.geolocationController.tracking = this.configService.getConfig('geolocate.activateDefault');
+      }
     });
 
     this.onSettingsChange$.subscribe(() => {
@@ -487,6 +505,70 @@ export class PortalComponent implements OnInit, OnDestroy {
       ).subscribe((sidenavMediaAndOrientation: [boolean, string, string]) => {
         this.computeToastPanelOffsetX();
       });
+
+    if (this.configService.getConfig('importExport')) {
+      const configFileToGeoDBService = this.configService.getConfig('importExport.configFileToGeoDBService');
+      if (configFileToGeoDBService) {
+        this.configFileToGeoDBService.load(configFileToGeoDBService);
+      }
+    }
+    // this.initSW();
+  }
+
+  private initSW() {
+    const dataDownload = this.configService.getConfig('pwa.dataDownload');
+    if ('serviceWorker' in navigator && dataDownload) {
+      let downloadMessage;
+      let currentVersion;
+      const dataLoadSource = this.storageService.get('dataLoadSource');
+      navigator.serviceWorker.ready.then((registration) => {
+        console.log('Service Worker Ready');
+        this.http.get('ngsw.json').pipe(
+          concatMap((ngsw: any) => {
+            const datas$ = [];
+            let hasDataInDataDir: boolean = false;
+            if (ngsw) {
+              // IF FILE NOT IN THIS LIST... DELETE?
+              currentVersion = ngsw.appData.version;
+              const cachedDataVersion = this.storageService.get('cachedDataVersion');
+              if (currentVersion !== cachedDataVersion && dataLoadSource === 'pending') {
+                this.pwaService.updates.checkForUpdate();
+              }
+              if (dataLoadSource === 'newVersion' || !dataLoadSource) {
+                ((ngsw as any).assetGroups as any).map((assetGroup) => {
+                  if (assetGroup.name === 'contexts') {
+                    const elemToDownload = assetGroup.urls.concat(assetGroup.files).filter(f => f);
+                    elemToDownload.map((url, i) => datas$.push(this.http.get(url).pipe(delay(750))));
+                  }
+                });
+                if (hasDataInDataDir) {
+                  const message = this.languageService.translate.instant('pwa.data-download-start');
+                  downloadMessage = this.messageService
+                    .info(message, undefined, { disableTimeOut: true, progressBar: false, closeButton: true, tapToDismiss: false });
+                  this.storageService.set('cachedDataVersion', currentVersion);
+                }
+                return zip(...datas$);
+              }
+
+            }
+            return zip(...datas$);
+          })
+        )
+          .pipe(delay(1000))
+          .subscribe(() => {
+            if (downloadMessage) {
+              this.messageService.remove((downloadMessage as any).toastId);
+              const message = this.languageService.translate.instant('pwa.data-download-completed');
+              this.messageService.success(message, undefined, { timeOut: 40000 });
+              if (currentVersion) {
+                this.storageService.set('dataLoadSource', 'pending');
+                this.storageService.set('cachedDataVersion', currentVersion);
+              }
+            }
+          });
+
+      });
+    }
   }
 
   setToastPanelHtmlDisplay(value) {
@@ -1058,6 +1140,13 @@ export class PortalComponent implements OnInit, OnDestroy {
     });
   }
 
+  private readLanguageParam(params) {
+    if (params['lang']) {
+      this.authService.languageForce = true;
+      this.languageService.setLanguage(params['lang']);
+    }
+  }
+
   private computeZoomToExtent() {
     if (this.routeParams['zoomExtent']) {
       const extentParams = this.routeParams['zoomExtent'].split(',');
@@ -1086,7 +1175,7 @@ export class PortalComponent implements OnInit, OnDestroy {
       const entities$$ = this.searchStore.stateView.all$()
         .pipe(
           skipWhile((entities) => entities.length === 0),
-          debounceTime(500),
+          debounceTime(1000),
           take(1)
         )
         .subscribe((entities) => {
@@ -1268,6 +1357,9 @@ export class PortalComponent implements OnInit, OnDestroy {
         undefined;
       if (version) {
         url = url.replace('VERSION=' + version, '').replace('version=' + version, '');
+      }
+      if (url.endsWith('?')) {
+        url = url.substring(0, url.length - 1);
       }
 
       const currentLayersByService = this.extractLayersByService(
@@ -1500,4 +1592,22 @@ export class PortalComponent implements OnInit, OnDestroy {
   onFilterSelection(event: object) {
     this.simpleFiltersValue$.next(event);
   }
+  zoomToSelectedFeatureWks() {
+    let format = new olFormatGeoJSON();
+    const featuresSelected = this.workspaceState.workspaceSelection.map(rec => (rec.entity as Feature));
+    if (featuresSelected.length === 0) {
+      return;
+    }
+    const olFeaturesSelected = [];
+    for (const feat of featuresSelected) {
+      let localOlFeature = format.readFeature(feat,
+        {
+          dataProjection: feat.projection,
+          featureProjection: this.map.projection
+        });
+        olFeaturesSelected.push(localOlFeature);
+    }
+    moveToOlFeatures(this.map, olFeaturesSelected, FeatureMotion.Zoom);
+  }
+
 }